Added functions for initial spawning.
When the FishManager starts, the start function makes a population data object for each fish type, then adds that to the populations dictionary. PerformIntialSpawning is then called which will spawn an initialCount (half the target population) amount of fish for each FishType distributed across suitable regions. Currently suitable region just checks if the speciesID of the region matches with the fish type, however camera postion/lure positon and other factors can determine if the region is suitable. Also the SpawnFishGroup creates a schoolParent, then as a child to that schoolParent, populates count amount of fish of fishType, then adds what was addded to the populations entry for the species. Also added OnGUI to draw popultion stats from the populations dicionary.