Made fishing ACTUALLY fun
There is now skill-based mechanics in fishing, tap the down arrow to release tension when reeling in a fish. Don't get ahead of yourself, if you tap too quickly (0.6s) it will ADD tension, increasing the likely-hood that the tension bar is expended and the fish breaks free. The fish also dont conglomerate in one spot anymore. Sprite was implemented instead of the triangle. Casting sets limit on x when hitting the water, making the fishing process way more realistic in terms of how you can maneauver the lure.